Use `Attribute::Simple` for simple attributes
Summary:
I've been looking into implementing attribute narrowing with a new type, which
also requires changes to the attribute resolution logic and the intermediate
representation in the `Attribute` type.

I realized that with the current `ReadOnly` and `ReadWrite` variants, I'd have
to duplicate handling (or worse yet make it inconsistent).

In addition to that forward-looking concern, the longer I looked at our code the
more skeptical I became that `ReadOnly` vs `ReadWrite` is a good distinction to
make: it forces downstream logic to decide whether to handle ReadOnly or ReadWrite
in cases where it really doesn't matter, and thereby couples the downstream
code unnecessarily to possibly-buggy upstream behavior.

A glaring example of this: enum value handling hardcodes that enum values must
be ReadWrite. But they really should't be ReadWrite, the only reason they are is
because no one had really noticed this behavior. If we tried to fix the upstream
bug, though, the match in `attr.rs` would break.

Given the combination of fishy existing behavior and that attribute narrowing
is less convenient with the current form, I'm changing the representation.

This diff is a simple mechanical change, the next couple diffs will simplify a few
cases where the new form allows for nicer matching

Reviewed By: rchen152

Differential Revision: D72012082

fbshipit-source-id: 9e80d4cf9ff0d7a98cd810c12a1f02ba49c076fd