Factor out `ReadableState` and `Transaction`
Summary:
In this stack, we will make `State` a transactional database, so that we can have concurrent recheck and read operations (mostly IDE operations). It will require us to split the current `State` into `State` and `Transaction`.

This diff does two refactors to first shape the future API to make the next diff smaller, but doesn't fundamentally change the implementation and data layout.

1. Factor out a bunch of readonly operations on `State` into `ReadableState`.
2. Rename `State` to `Transaction`, and create a new `State` that mostly wraps transaction. This prepares the future diff where `State` and `Transaction` are actually split in two, and most of the `State` access has to be done through `State.getTransaction`.

Reviewed By: ndmitchell

Differential Revision: D71939236

fbshipit-source-id: 2d72daf3c8c70c69dbb4665e727d670866faabf2