check whether metaclass call return type is unannotated
Summary:
I had to plumb this all the way from where we have access to the function definition AST in the binding phase.

This flag is weirdly specific (it's only true if the field is a function definition & the function definition does not have an explicit return annotation), but it's what we need here.

If anyone has a better name suggestion or an alternative place to put this, LMK

Reviewed By: ndmitchell

Differential Revision: D72024533

fbshipit-source-id: 741674974d0d5029e53ec340a253c6217aa136f0