Record we have bugs in loop inference
Summary:
When we are infering the types of these variables in the loop, we use a Var to capture everything. That works great for things that are constructive, like `x = [x]`. But when you do:

```
x = 100
while cond():
   x += 1
```

We need to know the type of `x` to resolve the `+=`, so we default it to Any, which gives incorrect types. The solution implemented over the next 5 diffs is to instead default to whatever the type was before the loop, if we do have to default to anything.

Reviewed By: yangdanny97

Differential Revision: D72046148

fbshipit-source-id: 49b82b862fd91a09bfc370e456e75d0e4e61ff1c