Refactor Pyrefly so fields aren't available on TypedDict
Summary:
We want to extract the fields from the `TypedDict` data structure - a few times we've run
into performance issues with it, most recently in the case where there are cycles of typed
dicts that refer to one another.

---------

The standard solution to this problem - similar to how we separated class *definitions* from
class *metadata* so that the definitions never point at one another directly and create
cycles - is to defer the potentially-cyclic computation through a computation in
`AnswersSolver` (we could do this using a binding, but that's not strictly needed and in
fact I opt not to in this stack). By doing so, we will ensure that the fields of one
typed dict never forces the fields of another typed dict, breaking the cycle.

Note that this is also probably better for perf, particularly if there are large "hot" typed
dicts, because we can compute the fields exactly once, caching it in an addressable Binding
that has a corresponding Idx, rather than having to re-create it from the Class repeatedly.

That is what this stack will do.

--------

This diff makes no actual structural changes to Pyrefly, but re-wires the API
to prepare for extracting fields, by making all functions that depend on the
fields part of `AnswersSolver` (so that they can
perform a Calculation to solve a Binding) instead of

Reviewed By: ndmitchell

Differential Revision: D72074622

fbshipit-source-id: ed62370f1fc554af664e67699fbb729a250e3d9f