Move and rename the typed dict fields calculation
Summary:
At a pure naming level, it was pretty confusing to have a function called
`get_typed_dict_fields` that returned not the fields themselves (like
`typed_dict_fields` does) but rather just the map that goes in metadata.

In addition, it was not very modular to have this function, whose only
legitimate use case is inside the *calculation* of class metadata (in a solve,
not downstream), be public.

Instead, I move it to `class_metadata.rs`, make it private, and rename it to
more clearly indicate it is calculating (from scratch) the fields for use
inside a `TypedDictMetadata`.

Reviewed By: ndmitchell

Differential Revision: D72093181

fbshipit-source-id: f28831ed7b5240701183ecf37efb9ca261d18371