Implement sanitization of fields that use method-scoped typevars
Summary:
This commit implements the logic to strip method-scoped type vars from fields whose type
is inferred based on method implementations, by falling back to a gradual type.

I don't yet produce an error message, that will be in the next diff.

The gradual fallback hook is currently best-effort given Pyrefly's current data representation,
which doesn't attach enough information to a Quantified for me to produce a non-Any type
when there's a bound on the type variable. I left a TODO to revisit this, since I suspect we'll
need those bounds for other use cases anyway, but it's pretty low-priority to give the best
possible inference in this edge case.

Reviewed By: rchen152

Differential Revision: D71906532

fbshipit-source-id: 5ae41f0d99fbf05b66b96e2a39436b419af4e1f1