Add a test case for another leaked Quantified
Summary:
I was discussing why typevar fields are tricky to detect and why my
first implementation (which tried to check that all quantifieds are bound)
did not work (because the quantifieds can be bound by a nesting function).

And I realized that the solution I found to that problem - which was to
specifically detect the use of method generics - leads to a different
issue: we still leak type variables in generic classes that are nested
inside other generic classes.

The solution to this issue isn't obvious, because the semantics may not
even be well-defined. As a result I'm leaving a TODO here and not planning
on working on it in the near term.

Reviewed By: rchen152

Differential Revision: D71995896

fbshipit-source-id: a0c363b8bda132545753b0b24792d811e5bfc9c5