Test case for refining attribute type
Summary:
I was examining this paste P1769351799, which contains 800+ occurrences of errors of the form : Object of class ... has no attribute `type`.

Neither mypy nor pyright fail on this testcase.

mypy just refines n.type to be int and the addition succeeds.

pyright lets n.type be any and allows any + int to succeed.

We are considering the "None" part of the optional annotation and erroring on that.

I am not sure exactly how many PyTorch attribute errors are happening due to this specific case, but we are more conservative than mypy and pyright (on strict mode). Is there a world where we can consider admitting a program like that?

Reviewed By: stroxler

Differential Revision: D72078630

fbshipit-source-id: 3f97c8da5588fe5d25c1c9611756c7e47b9d92b8