Return error if config search path or site package paths don't exist
Summary:
Addresses https://github.com/facebook/pyrefly/issues/59

We've run into issues before where we think we're type checking using a site package path, but it actually isn't being used. This diff adds validation and prints out an error message when we see that we're not using an entry in `search_path` or `site_package_path`.

There's a question here around whether this should just be a warning, or if we should actually error and refuse to type check when this happens. I think a warning is fine for now, since we may run into the same issues that Pytorch and other OSS projects have with some dependencies not being installed.

Reviewed By: grievejia

Differential Revision: D72078439

fbshipit-source-id: ecd00d1979e7206e39233998063bd296497fc0a5