Make `State` a transactional database
Summary:
We do not want to block all IDE services while an check is ongoing, because doing so will be very bad for IDE request latencies. Right now, this is inevitable, because the part of `State` that's suitable for read is shared with the part of the `State` that's only useful for an ongoing check. In this diff, I will separate `State` into two parts: the main `State` that represents a snapshot after a recheck, and a `Transaction` that contains temporary structures and updated information during checking.

Under the new system, we will need to have a snapshot of `ModuleData` living in `ReadableState`, some of them can be copied into `Transaction` during checking as `ArcId<ModuleData>`, and committed back to `State` all at once after the check is done. For now, we will make a copy from `ModuleDataSnapshot` in the main state into `ArcId<ModuleData>` whenever we want a `ArcId<ModuleData>` during checking, even if it doesn't plan to mutate it. While we can implement a fancier copy-on-write mechanism, benchmark shows that this diff barely has any noticeable effect on perf.

Finally, to ensure that recheck is well-coordinated so that we don't run into bad situations, we use `State` to make sure that the following holds:
1. There can be at most one ongoing recheck that can eventually commit, because we don't really have a way to merge together the results of two arbitrary rechecks.
2. All the reads over the state are reads over a consistent view  (i.e. it won't observe a mix of state between different epochs), which is enforced by
    a. There can be as many concurrent reads over state as possible, but they will block committing.
    b. During the committing of `Transaction`, all reads will be blocked.

Reviewed By: ndmitchell

Differential Revision: D71357622

fbshipit-source-id: 3ad92c84493050c99d92f6f6326b35af61a9f2b1