No need to use double references
Summary:
I was staring at this figuring out what it does (as part of understanding API
changes needed for attribute narrowing) and realized the double `&&` in inlay
type hints were irritating me.

It's simpler and in principle might be the tiniest bit faster (depends on the
compiler, probably) to avoid chained references.

Reviewed By: connernilsen

Differential Revision: D72251880

fbshipit-source-id: 6fa94c28f3df4e453e532f48f0cdf34a70f5bb8f