Don't clone types just because we are checking
Summary:
There's an inefficiency baked into pretty low-level Pyrefly operations right
now: because the `check_type` function's historical origin is that it was a helper
in contextual typing (so that `expr` could be built on top of `expr_infer`, it was
always designed to check a type and produce a new type at the same time.

But in fact at this point it is *extremley* common to check a type without needing
to produce one - function calls in particular need this, but so do many control
flow and operational models that require validating types which don't actually
propagate anywhere.

We were paying for clones on all of these operations... but it's even worse
than that, because even in the cases where we *do* need a Type to propagate, we
almost always already have an owned copy in the caller - we were passing that
owned copy by reference and then cloning it in the happy case! This is pretty
inefficient; it's probably fine to clone `want`, which (a) is frequently not
owned because it's typically being pulled out of some other context and (b)
isn't cloned anyway except in the case of a type error, which should be at
least a couple orders of magnitude less common than the no-error case.

The fix is to tease apart the check, which should be by reference, from the
override-with-`want` behavior which can and should take `got` by value (but
can clone `want` as needed).

This change will make it easier for me to make the attribute narrowing refactors
I need without raising questions around efficiency, but I actually think it might
wind up being a nontrivial perf win by itself, these clones are *super* common.
It will be intresting to keep track of our benchmarks.

Reviewed By: ndmitchell, grievejia

Differential Revision: D72257345

fbshipit-source-id: ce1c661c26668bba4802c9a18ed028692f60d26a