Demonstrate a bug in tuple unpack handling.
Summary:
We abort type checking if we hit a non-unpackable unpack. This is not good
behavior, because the user could easily suppress a false positive on the bad
unpack, and we wouldn't warn them about code in the tail of the tuple that's
clearly going to crash the program.

This bug is important to track so that I can properly enjoy pontificating about
structured programming in my next diff summary.

Reviewed By: migeed-z

Differential Revision: D72262805

fbshipit-source-id: 367373b31dd90ce43770ca8a35c534bd38079955