Create `BindingExpect::TypeCheckExpr`
Summary:
In order to work on attribute narrowing, I have to manually audit every single
use of expressions in binding logic to understand which ones need attribute
narrowing.

One issue I ran across is that there are a few cases where an `Binding::Expr`
is used exclusively to ensure an expression gets type checked, but there is no
downstream use.

It would be nice to signal the intent here, because it's not always obvious
whether or not the type might get used downstream (for example, I thought
expression statements were probably an example of this, but actually they are
needed for handling `Never` return types correctly, which is not something you
can understand from looking at any one block of code).

At this point the only case where I'm 100% sure we don't use the result downstream
is the message on assert statements, but I suspect there are more and I can cut
them over as I find them.

Reviewed By: ndmitchell

Differential Revision: D72322944

fbshipit-source-id: 870e2b7e7ad149aa697a84fce9a1e97d30d90297