Separate BindingExport from Binding
Summary:
I'm fairly confident at this point that attribute narrowing will be handled
by having bindings operate on some kind of `TypeInfo` that represents both
a type and potentially a tree of attribute narrows active in the current flow.

But we don't want exports to resolve to attribute narrows, even if there are
some active in the flow at the end of a module: we want only the bare type.

This means we really need `BindingExport` to be a separate entity from `Binding`;
this is pretty easy to prep for by just wrapping the underlying Binding in
a new type; when we introduce attribute narrowing we'll strip any narrows
and just extract the `Type`.

It's probably better for clarity anyway: the keys and bindings aren't always found
right next to one another in the code, and with this change it's a lot easier to do
a quick search and find all the export bindings we make, which is a useful thing
for maintenance.

Reviewed By: ndmitchell

Differential Revision: D72319607

fbshipit-source-id: 0d3cda330937e55091dcd45bc69771e5b19717ca