Don't shadow a very similar but parallel name
Summary:
The `function_def` function is pretty big, which means good style hygiene is
more important than in a small function.

We have two different keys that track the "previous definition of the same function
name", which has to be handled very carefully because normally definitions just
shadow existing values but for `overload` they have to be combined: `pred_idx`
is the previous definition at the "value" level (i.e. a `Key`) whereas
`pred_function_idx` is the same previous definition at the "function def level"
(i.e. a `KeyFunction`). Both are defined at the very top of `function_def`

What's confusing (I ran into it when exploring attribute narrowing changes, which
potentially change the semantics of `Key` slightly) is that we have an `if let`
block where the unwrapped `pred_function_idx` shadows `pred_idx`.

This is weird, and hard to follow because the defs are hundreds of lines apart.

Let's just rename: an `if let Some(_) = _` probably should just shadow the same
name on both sides.

Reviewed By: rchen152

Differential Revision: D72349526

fbshipit-source-id: 7aebda2cc2389ae756bac0f371496e7bc17d5aa7