store QuantifiedInfo in Quantified
Summary:
This makes Quantified not copy-able anymore, so we have to clone in a bunch of places.

Due to an accidental amend that I didn't realize till it was too late, this diff also includes an example of checking a type parameter's bounds/constraints during binding solving.

There's a few other rules like default <: bound, default == any constraint, and some more rules around what defaults are permissible for TypeVarTuple and ParamSpec, which I'll implement later.

Reviewed By: rchen152

Differential Revision: D72267091

fbshipit-source-id: 1a0d8f1cdf1dc3fb9593a56a7d33346101bb34fa