Add python_interpreter to config
Summary:
# This diff

This diff adds the `python_interpreter` config and flag options. This will be an executable that we can use in the next diff to determine `PythonEnvironment` values by calling it and passing in a script to run.

# The next few diffs

The next few diffs will add functionality for trying to determine some values for type checking from the system Python installation if not provided by flags or a config instead of falling back to the default. This is a big UX improvement that will match whatever settings are used by the Python interpreter while executing.

This will be done by making these values optional, and only falling  back to Pyrefly's default values if they're still unset when they're needed. We will try to query (and cache) these `PythonEnvironment` results for use in all the configs we load and use them when possible in loaded configs with unset values. On failure, we fall back to Pyrefly's defaults.

## Upcoming diffs
1. Make future `PythonEnvironment` values optional
2. Create a the `PythonEnvironment`
3. Add `python_interpreter` to `ConfigFile` (this diff)
3. Add functionality to query `python_interpreter` for values in the `PythonEnvironment`
4. Add a function to update the `ConfigFile` with a `PythonEnvironment` so the values are set

Reviewed By: ndmitchell

Differential Revision: D72093200

fbshipit-source-id: 1e86b6b059321bca62a66c9d0cecc9ef537bf9c8