add auto environment configuration
Summary:
# This diff

This diff adds caching for `PythonEnvironment`s. We do this by declaring a global (mutex-protected) map, which we can use to look up any values we've already found.

In the case a query fails, we emit an error message (one time), and then set the value to `None`. If we look up a value, and it's none, then it must have failed previously, so let's not try to recompute it. If a user changes their interpreter, they must restart to update here.

For switching between conda environments, we can add some logic down the line to handle the change (unless we're planning on just restarting for that).

# The next few diffs

The next few diffs will add functionality for trying to determine some values for type checking from the system Python installation if not provided by flags or a config instead of falling back to the default. This is a big UX improvement that will match whatever settings are used by the Python interpreter while executing.

This will be done by making these values optional, and only falling  back to Pyrefly's default values if they're still unset when they're needed. We will try to query (and cache) these `PythonEnvironment` results for use in all the configs we load and use them when possible in loaded configs with unset values. On failure, we fall back to Pyrefly's defaults.

## Upcoming diffs
1. Make future `PythonEnvironment` values optional
2. Create a the `PythonEnvironment`
3. Add `python_interpreter` to `ConfigFile`
3. Add functionality to query `python_interpreter` for values in the `PythonEnvironment`
4. Add a function to update the `ConfigFile` with a `PythonEnvironment` so the values are set
5. Cache `PythonEnvironment`s (this diff)

Reviewed By: grievejia

Differential Revision: D72091005

fbshipit-source-id: 88b7febab556d420341dec50c59568abf2e2ce83