Make python environment values optional in ConfigFile
Summary:
# This diff

This diff prepares for upcoming changes by making `python_version`, `python_platform`, and `site_package_path` options. There should be no functional changes in this diff.

We make them optional here so that we can determine if they're set by a config file (or flag), and if not, try to use the system Python installation (or another interpreter) to get the values. If no value is set, then we can fall back to Pyrefly's defaults.

# The next few diffs

The next few diffs will add functionality for trying to determine some values for type checking from the system Python installation if not provided by flags or a config instead of falling back to the default. This is a big UX improvement that will match whatever settings are used by the Python interpreter while executing.

This will be done by making these values optional, and only falling  back to Pyrefly's default values if they're still unset when they're needed. We will try to query (and cache) these `PythonEnvironment` results for use in all the configs we load and use them when possible in loaded configs with unset values. On failure, we fall back to Pyrefly's defaults.

## Upcoming diffs
1. Make future `PythonEnvironment` values optional (this diff)
2. Create a the `PythonEnvironment`
3. Add `python_interpreter` to `ConfigFile`
4. Add functionality to query Python for values in the `PythonEnvironment`
5. Add a function to update the `ConfigFile` with a `PythonEnvironment` so the values are set

Reviewed By: ndmitchell

Differential Revision: D72091678

fbshipit-source-id: c2992b55f0bd711edf27e829c76fe8f9f33947a8