Add functionality to query Python interpreters for environment information
Summary:
# This diff

This diff adds a method on `PythonEnvironment` that allows us to query a given Python interpreter for information about it. Specifically, we're looking for the Python version, platform/os, and site package path. We could figure out the Python version and platform in other ways, but might as well let Python tell us what's going on to keep things as closely aligned with it as possible, especially if we're already starting a subprocess to query a Python interpreter for some of this information.

# The next few diffs

The next few diffs will add functionality for trying to determine some values for type checking from the system Python installation if not provided by flags or a config instead of falling back to the default. This is a big UX improvement that will match whatever settings are used by the Python interpreter while executing.

This will be done by making these values optional, and only falling  back to Pyrefly's default values if they're still unset when they're needed. We will try to query (and cache) these `PythonEnvironment` results for use in all the configs we load and use them when possible in loaded configs with unset values. On failure, we fall back to Pyrefly's defaults.

## Upcoming diffs
1. Make future `PythonEnvironment` values optional
2. Create a the `PythonEnvironment`
3. Add `python_interpreter` to `ConfigFile`
3. Add functionality to query `python_interpreter` for values in the `PythonEnvironment` (this diff)
4. Add a function to update the `ConfigFile` with a `PythonEnvironment` so the values are set
5. Cache `PythonEnvironment`s

Reviewed By: grievejia

Differential Revision: D72202641

fbshipit-source-id: a32e0d6c3bc58b2e52c64b4ffc397905a697b918