Add unit tests emulating nondeterministic
Summary:
Jia and I had a discussion at lunch about nondeterminism in Pyrefly, and I
explained my pet theory which is that it's caused by us being too lazy about
export solving.

Here's what I believe to be the case:
- Pyrefly tries to be as lazy as possible, so if an Export gets looked up,
  we traverse the binding graph for it as sparsely as possible.
- Any Expr forced anywhere in the binding graph is capable of mutating
  the Vars, because the solver will eagerly pin them
- The order in which exports get looked up can never be deterministic,
  because it depends on other modules which we obviously have to do
	in parallel

The emulation demonstrates the root cause using two unit tests - they
import the same module, but they force the exports in different orders, which
causes the type of `x` to be resolved differently.

You can't see it in the test itself, but if you actually run the tests it
will also print out the type errors for the `nondeterminitic` module, and
you'll see different type errors in unit test a vs unit test b; in one case
it complains about the string append and in the other about the int append.

Talking with folks, I think some people thought this couldn't happen because
we force `Var` to `Any` in exports so they can't leak, but that's incorrect - we
are only preventing a `Var` from leaking directly through a return value, we are
not preventing side-effects of export resolution from pinning `Var`s in
nondetermistic ways.

------

My best guess at how to fix this: solve for all exports when we solve for
any export. The sparse traversal would still happen, but in a deterministic
order.

This would come with some perf cost, although I think it should be pretty
modest on most library modules; unless the module does a lot of top-level
control flow, forcing just the exports should usually be pretty cheap.

Reviewed By: ndmitchell

Differential Revision: D72412153

fbshipit-source-id: c62b7ff57eddf2126d59a998ba765e9e5b5fd164