Thread TypeInfo through expr logic
Summary:
The heart of the refactor to use `TypeInfo` that can track attribute narrowing
is not actually binding - only a handful of binding cases actually need
TypeInfo, most of them encode control flow operations that are type-level only.

As a result I decided it makes sense to start the refactor by threading `TypeInfo`
through the `expr` methods first, which is also where the most inconvenience
happens becasue *most* expressions encode operations that produce new values and
threfore are also Type-level.

The solution I came up with is this:
- First, all existing `Type`-level interfaces in the public API must be preserved;
  without this, the existing `Type`-level analysis throughout Pyrefly would be
	very difficult to update since we'd have to start unwrapping the inner type all
	over the place.
- Most of the existing public APIs, however, need to built in terms of `TypeInfo`
  because we need `TypeInfo` versions of some of the top-most apis in order to
	successfully bind top-level names to attribute-narrowed type info.
- The implementation of `expr_infer_type_info_with_hint` is *mostly* type level; we
  already had it structured to produce a local variable `ty` that is used for
	recording traces for IDE support, so I just take the same variable and wrap it in
	a `TypeInfo`.
- The three cases I know of where v0 atribute narrowing needs to propagate attributes,
  which are `Name`, `Named`, and `Attribute`, use a `return` to break out of the
	block creating `ty`; this is necessary because the code below that block is always
	producing nullary attribute refinments, but these three cases are going to need
	to explicitly handle them.

Note: I improve on this first approach in the next diff

Reviewed By: ndmitchell

Differential Revision: D72319610

fbshipit-source-id: 918fa8a5faa4deeac5e8041031bb59e890ab0250