Optimize expr_infer a bit
Summary:
In the preceding diff, I brute-forced the work of making the
`expr_infer` logic able to handle `TypeInfo` at the bottommost
layer (which is needed because attribute narrowing will bubble
out from `Name` resolution).

But it comes at a cost - in all but three expression cases, we
are really operating at the type level, which means the preceding
diff potentially did quite a lot of wrapping and unwrapping
on complex expressions.

This diff cleans things up by shuffling code between the two
functions `expr_infer_type_info_with_hint` and `expr_infer_with_hint`
so that the latter is `Type`-level only:
- `expr_infer_type_info_with_hint` only matches on the cases that
  care about attribute narrowing, and defers the work to
	`expr_infer_with_hint` for all other cases, just wrapping the result
- `expr_infer_with_hint` operates entirely at the `Type` level, and
  deals with the cases that preserve narrowing by just getting the type
	info and unwrapping the inner type

This should perform a bit better, and I think it's actually a little
clearer in the long run too since we have two matches that refer to
one another and give us a clear place to comment with what's going on.

It would be possible to optimize a bit more by making a "don't preserve
attribute narrowing" version of the cases that do preserve narrowing to
avoid extra work and clones but that would be harder to do without making
the code more complex so I think we should defer it for now.

Reviewed By: ndmitchell

Differential Revision: D72319606

fbshipit-source-id: eb68d893ca74bedfe7656c7578b7fee73825f02c