Replace any ignored module with Any always
Summary:
Zeina found that mypy's behavior of `ignore-missing-imports` *always* sets an ignored module to `Any`, even if it's able to be found.

While that behavior is quite different from what we expect, I think matching it would be a good idea, especially since the projects she's testing on have other errors that arise from this.

I think that because a project may/may not have that functionality available, it's probably best right now to copy that. I'm going to rename this config in an upcoming diff to something like `rewrite_import_as_any`.

Reviewed By: migeed-z

Differential Revision: D72432739

fbshipit-source-id: 664b9bd44b94fe5a989430e9942a6a013a363c8e