initial support for vscode workspace folders
Summary:
The IDE needs good defaults when there is no Pyre2 config. One option for these defaults is the IDE: Workspace folders are easy ways to learn roots and the VSCode interpreter setting is a good resource for determining which interpreter to use. If any of these options are overriden in a Pyre configuration, we should use the option in the config. In their absence, we should use these items from VSCode.

This diff starts using the workspaceFolders capability in IDEs: We create a "`Config`" for each root and additionally a default `Config` for files not covered by any config. When responding to IDE requests, we first find the config the file is referencing (it should only reference one), then perform that request under that `Config`. This is a temporary solution until connernilsen implements configs fully. When implemented fully, pyre configs will supercede these IDE configs in order for CLI/IDE to match. Currently, every `Config` gets any search_paths and site_packages referenced in the command-line. This is subject to change once we no longer need the command-line override.

Next diffs:
- add workspace folder root as search_path
- support didChangeConfiguration to handle workspace folder updates in VSCode (`Config` will likely need to change to `Arc<Mutex<Config>>` since will become mutable)
- send pythonpath to language server as part of didChangeConfiguration in order to support interpreter settings

Reviewed By: SamChou19815, connernilsen

Differential Revision: D72419050

fbshipit-source-id: 435b02d124e04ae182d248e9d5e6544d3d023bba