Wire up the Key/Binding Result as TypeInfo
Summary:
This diff threads the `TypeInfo` type, which at the moment is just a wrapper
around `Type` but has the ability to be enriched with other static analysis
information (which we will use to track attribute narrows), through the
Answers table.

Something of note in this diff is that the (Rust) type information really isn't
ideal - there are a lot of Binding variants that really represent an operation
that creates a new value (for example, computing the type of an iteration value
given the iterable), and these really ought to evaluate to a `Type`.

But after trying to tease these apart and running into issues with how `Var`s
get resolved, I concluded that now is not a good time to try to deal with that
because the table macro loop order changes if we try to split the existing
`Key` type, which causes all sorts of observable behavior changes in edge cases.

As a result, I strongly believe that to move forward with attribute narrowing
we should just eat the annoyance of having to always unwrap types, and we can
figure out how to represent things better later... this is especially true
given that we have some issues with determinism already, so trying to re-wire
things around when exactly Vars get forced might be premature.

Reviewed By: ndmitchell

Differential Revision: D72361203

fbshipit-source-id: 0682a06dadaef73202b9a9a17948116f662506a1