Add plumbing for a config to skip analyzing functions
Summary:
We would like the option to disable:

1- analyzing the function body
2- inferring the return type

This is only for unannotated functions though.

Similar to pyright and mypy, we think that creating a single config/flag combining both points would be a good direction.

Note: having this config for the whole project is a good idea for PyTorch, because it will act exactly how `allow_untyped_defs` is on mypy. Right now, all unannotated functions are not analyzed, and this would enable us to do the same.

We could consider configurations per directory and per file as next steps, but this seems like a good first step and will do what we need for PyTorch.

Right now, I'm working on adding a flag for this and my goal is to find a way to thread this flag to the typechecker and skip analyzing functions when the flag is disabeled.

Reviewed By: connernilsen

Differential Revision: D72432085

fbshipit-source-id: 31355710a16e6bf48e5a666d634d1405bb46c1e6