Adjust error count display
Summary:
Previously we only print total error count and suppressed+disabled error count. This is fine but not 100% desirable because total error count is a bit of a misleading number. What's really matter to us (and what's being tracked on the IG dashboard) is the number of errors that are shown to the users, and currently there's no way to see it other than `wc -l` on the error output.

This diff proposes that we change the stats printing so it clearly shows
- How many errors are shown to the user.
- How many errors are disabled via config options
- How many errors are suppressed with inline comments.

Reviewed By: stroxler

Differential Revision: D72486040

fbshipit-source-id: da13284550bc33a42b9086f3b4000da764ee47c2