Display the attribute narrows for `TypeInfo`
Summary:
If we call `reveal_type` on a name with active attribute narrows, we'll see
the narrows in parenthesis.

I'm pretty sure this will be useful for users (certainly if Pyre1 had this
it would make some oncall questions easier to answer!) but the most immediate
use case is that, as I start implementing the actual logic for narrowing,
this will help me make it testable before the logic is complete enough
to get type-level behavior.

There's a test but only of the trivial case; I can't add a test with
actual narrows yet because I need to build the API for attaching a new
narrow; I'll do that in the next diff.

Reviewed By: ndmitchell

Differential Revision: D72496979

fbshipit-source-id: 8268fdf92666121add80baef0d7caabe71a8a7f7