Partial API to add narrows, extended display test
Summary:
This diff couples an initial API for mutating `TypeInfo` to add attribute narrows with
an extended test of display; together the creation and display allow us to unit
test building and showing a TypeInfo with actual attribute narrows.

The API is good enough to test so I'm stopping here to keep the diff small, but there's
a case we aren't handling yet when we narrow directly on top of a `Leaf` narrow; the
new subtree is wiping out the leaf.

---------------------

Note: I need to tweak the API to handle that next, there's a TODO where the implementation
has a gap and another in the test where the problem shows up; it's not super trivial
to fix and most of the code in this diff is correct, which is why I'm breaking the work
into two diffs.

Reviewed By: ndmitchell

Differential Revision: D72496972

fbshipit-source-id: 78673fc3d5f8d63d325e4a0037f0556773c883bc