Finish the API for adding narrows; handle overlaps
Summary:
In order to fix a bug noted in the preceding diff, we need to be able to merge a new
narrow into a `NarrowedAttr`, which was not working.

This is actually pretty tricky to do: I originally was trying to use a mutable
API, but because I did the static typing thing where instead of just using a
pair of options I used an enum I have to figure out how to wipe an enum variant
in-place.

My Rust-fu is not good enough to do this, every time I tried to do it (the main
option I saw was to try to get an owned handle on the data by derefing it, then
assign back to a deref and hope the compiler sees that this is okay) I failed.

After an hour or so of fighting I realized it makes no sense to hold up all of Pyright
development for an irrelevant detail (is this a mutable operation or a destructive one?).

One side effect of giving up on mutation for now is that in the higher-level code,
the only way I could work out to do a map merge is to pop the existing `attr`
off the map and re-insert a new one. This creates the visible-in-reveal-type behavior
where the first narrow added on `x` moves to the end of the map when we add a new
narrow on the same subtree... not an actual problem (the behavior is still deterministic)
but I figured I should note what's going on here.

----------------------

Advice on how to improve this is welcome, although I think it's good enough already
to continue working on the actual logic so I'll probably fix forward rather than
rewriting this code unless we discover something is horrifically wrong.

Reviewed By: ndmitchell

Differential Revision: D72496973

fbshipit-source-id: 0cb8f245fd6c29e2d3b1510bcf9160372d42a403