Expand TypeInfo
Summary:
Up until this diff, my entire attribute narrowing stack has been plumbing: putting
a layer in between bare `Type` and the result of name / expression resolution
so that we would have a place to store attribute narrowing information.

This diff begins the actual implementation of attribute narrowing by adding to
`TypeInfo` an optional rose-tree like structure of attribute narrows.

We don't yet actually create the attribute narrow tree, this diff is entirely

NOTE: Conceptually it would be cleaner to just directly store the `OrderedMap`
and have it be empty in the case of no narrows. But an empty map cost significantly
more RAM to store and compute to create than an option that is usually None - I
am pretty sure the approach here is going to be noticeably more efficient. Happy to
switch to a empty map if reviewers prefer it.

Reviewed By: ndmitchell

Differential Revision: D72496974

fbshipit-source-id: e1522fb86acdca26b8e37b2c4d2214417d1bea21