Don't force ParamSpec vars prematurely
Summary: test_paramspec_twice fails because we force a ParamSpec var before we have an answer for it. This diff deletes the premature force_var() call and instead passes the var to callable_infer_params. Subsequent diffs will implement matching arguments against the var (and fix the test that this diff breaks).

Reviewed By: stroxler

Differential Revision: D72504388

fbshipit-source-id: dbc1435192e6da7a9ebf3890f07e72551192ca9d