Demonstrate nondeterminism on arbirary calls
Summary:
After a discussion Friday about the problem of nondeterminism, I agreed
to take a look at it.

There's a general attitude I don't like though - that the problem here is
empty containers. That's obviously false at the level of architecture,
empty containers are just one more expression that produces a `Type`, and
the bug is caused by the solver forcing `Var`s.

It's true that empty containers are *one* example of producing dangling
`Var` types that can get forced in a nondeterministic order, but I'm concerned
that there are probably many other mechanisms by which `Var`s can leak
out (in a follow-up diff where I prevent `Binding` from leaking them, we see
that there are indeed more cases I hadn't thought of in advance like lambdas).

I wanted to demonstrate quickly that we don't even need a constructor call
to produce the dangling Var problem - all we need is a generic function where
the call fails to pin a type var. You can most easily do this using a
dangling return var, which I did.

But I wanted to make it clear - because again I see whack-a-mole-ism proposals
coming, Pyright checks for this pattern syntactically - that the same problem
can be created in ways that a symple syntactic check will *not* catch e.g.
if a parameter has a type var that might not get pinned by a specific argument
(which is very easy to trigger when you have a type generic over a union).

Again, this makes me want to think systematically about determinism instead
of trying to case-analyze our way out of the problem.

-----

Note that because we deep force types for both `assert_type` and `reveal_type`,
it's tricky to actually demonstrate the behaviors I need to highlight. I used
a hack with the `force_error` function which lets me observe the `@_` that
tells us we have problems.

-------

I also introduce a new module `calls.rs` to our tests because I realized there's
no obvious dumping ground for call-related unit tests right now, which has come
up a few times for me as I think about various edge cases.

Reviewed By: ndmitchell

Differential Revision: D72518939

fbshipit-source-id: 08874453236535bec4cd7fd843e6ac2f73cd622a