Add doc comments to `NarrowedAttr`
Summary:
A narrowed attribute is an optional Type with an optional subtree; this
is encoded in an enum to express the invariant that at least one of the two
has to be present (or else we wouldn't have a node in the first place).

FWIW we might simplify this to a pair of options at some point in the future,
I think the use of a more informative enum type might not really buy us that
much here (in particular, mutations are probably easier using a pair of options
so that the outer struct doesn't have to be destroyed).

But for now this is what we have, so adding some comments which is both generally
useful and was requested by review.

Reviewed By: ndmitchell

Differential Revision: D72513030

fbshipit-source-id: 5e6d5945ebc8dbc11211bdfc5df3805c53e4cd34