Tighten up the display test code
Summary:
Instead of creating two class types and cloning them a lot, it's cleaner
really to have a higher-level helper that always makes a fesh type
and is concise enough to inline so the clones aren't needed.

In the process, it's also reasonable to use a distinct type for every
slot in our tree.

Reviewed By: ndmitchell

Differential Revision: D72513029

fbshipit-source-id: 2603e807a3cde9461dea3bc5aa566376ca4da0fb