Add AnswersSolver::instantiate() method and better docstrings for specialize, promote, and instantiate
Summary:
I realized that Class:instance_type() doesn't take TypedDicts into account, so it can be used to accidentally create a ClassType of a TypedDict class. To properly handle TypedDicts, the instance_type() method needs to be on AnswersSolver instead, so it can check class metadata. This diff adds a new AnswersSolver::instantiate() method that is meant as a replacement for Class:instance_type() and uses it in instantiate_fresh(). I'll move over the remaining callers in subsequent diffs.

Also adds docstrings that clearly describe the differences between specialize, promote, and instantiate.

Reviewed By: ndmitchell

Differential Revision: D72521981

fbshipit-source-id: cfc3283111fde2362eac268a5df6cf2d443c0efc