Use Quantified::as_gradual_type to fix bug in create_default_targs
Summary:
We don't want to use the bound.

Note that this changes create_default_targs to always use AnyStyle::Implicit rather than AnyStyle::Error, even when a range is passed in. I think this is fine, given that we don't use Any at all when a TypeVar has a default, so it's not like we were consistently showing "Error" previously. This change doesn't affect our ability to log an error in a potential strict mode down the line.

Reviewed By: ndmitchell

Differential Revision: D72471707

fbshipit-source-id: 060bff38fc42c484633f7b7856a45c6c7f2f45e7