Use Quantified::as_gradual_type to fill in missing type arguments to a class
Summary:
The typing spec doesn't specify what to do here, so it would be fine to keep the existing behavior, but for consistency/simplicity, we might as well use as_gradual_type here.

I had to change the "empty" value for a ParamSpec to match the expectations of the conformance tests.

Reviewed By: ndmitchell

Differential Revision: D72521034

fbshipit-source-id: 0b5b10c566945f47419f8e06f998749711ed8a4a