Make constructing raw singleton NarrowOps into helper calls
Summary:
It has been consistently giving me issues, as I consider various refactors of
NarrowOps logic, that the `pattern.rs` code is super coupled to the precise
layout.

That's probably reason enough to clean things up a bit.

But there's another readability benefit I think, which is more manageable
levels of syntactic noise in any one place. These `smallmap!` macros are just
incredibly dense, because they mix nontrivial syntactic complexity from
creating `NarrowOps` (which is a relatively complicated data structure, a map
of tuples) with the syntactic complexity of producing the requisite bits
of data from the raw pattern expressions (which is not trivial either, we're
doing things like dereferencing and then cloning).

By splitting it, neither place in the new code is nearly as hard to skim as
what we had before.

Reviewed By: rchen152

Differential Revision: D72541156

fbshipit-source-id: 48fb614f86f50b114bbbf7f19fd9929341c8726c