Simplify NarrowOps name logic a bit (small cloning cost)
Summary:
I need to extend the narrowing target to be potentially not just a
bare name.

But the exsting code is awkwardly coupled to not actually producing
`Name` but instead returning references to raw Ruff AST nodes.

This does enable exactly one perf hack: in the case of operators,
which can be chained, the existing code avoids an extra clone of the
names by working with references for as long as possible.

But preserving that perf hack across the generalization to attribute
narrowing would be quite challenging and I doubt it matters that
much - I'm choosing to just simplify for now.

-----

If anyone objects to the extra clones introduced here, I'm happy put a TODO in
and to restore the perf hack once attribute narrowing works; we'll probably
just need two versions of the narrowing target data structure, one that owns
its data and one that doesn't and has a lifetime annotation. It's doable,
I'm just not convinced it's worth the complexity.

Reviewed By: ndmitchell

Differential Revision: D72496976

fbshipit-source-id: ce4b28d844c7b64d3f499645cbf0dbf4050fccd7