Split narrow ops into two levels
Summary:
After digging through the narrowing code in moderate detail, I realized
that the way we model control flow through boolean expressions is a bit
peculiar (we traverse the same expressions many times, narrowing repeatedly at
various scopes), but it does appear to work.

I don't think it's really feasible for me to change the way it works deeply
in order to enable attribute narrowing, which means I need to find a way to
inject attribute narrowing into approximately the existing architecture.

## A quick primer on narrowing in Pyrefly today

The most important feature of the current architecture is that instead of trying
to model control flow *through* control-flow expressions (boolean operators +
the ternary operator), we simply duplicate it by re-creating the expression
control flow inside the `NarrowOp` tree, and executing *both* variations of
control flow.

Consider an example like this, where there's a large expressin with multiple
narrowing sub-expressions:
```
if x is not None and f(x) and isinstance(x, Foo):
    g(x)
```

In expression resolution, we use bind operations that happen inside of
expression traversal (if you look in the `expr` binder, you'll see that we bind
narrow ops inside the cases for both Expr::BoolOp and Expr::If) to modify the
scope in the middle of an expression, so that narrowing can be accurately
modeled through the expression-level control flow. This is the mechansim by
which the narrow from `x is not None` propagates to the type check of `f(x)`,
and for that matter also to the type *checking* portion of `isinstance(x,
Foo)`.

But that expression-specific scope mangling actually has no effect whatsoever
on what we get at `g(x.y)`. Instead, a completely separate traversal of the
entire expression `x is not None and f(x) and isinstance(x, Foo)` goes in
and creates, from scratch, a `NarrowOps::And` instance that understands we
should do both an `_ is not None` narrow *and a `isinstance(_, Foo)` narrow,
in that order. This and node starts from scratch with the original type of
`x`, and re-produces the same narrowing described above as it proceeds through
the And; the final result is used for the body of the if.

Importantly, we then negate the full tree into a `NarrowOps::Or(negation, negation)`
structure that gets used for the `else`, if there is one.

## How this impacts attribute narrowing

The expression-traversal part of the existing infra is going to work just fine
almost no matter how we implement narrowing for attributes: as long as we are able
to combine root node narrows with attribute narrows, the narrowing will cascade
through the solve of bindings and it works.

I had originally thought this might be all we need, and that we could therefore
use separate bindings for attribute vs normal narrrows.

The problem is the other case: when a large expression with many sub-expressions
is narrowed, the entire thing is a single unit of narriowing and we sort of
"drop the initial type" through that entire unit (and potentially later its negation)
as a while.

And it is absolutely critical that this can operate on top-level and attribute narrows
at the same time: consider an example like
```
if x is not None and x.y is not None and x.y.z is not None
```
it's obviously critical that as we try to narrow these, we are able to evaluate the
left narrowings before applying the right ones (and understand the implications of
negating the whole thing).

So, we're going to need a tree that can express top-level and multiple attribute
narrows in a single `NarrowOp`.

This means the operations need to be taggable with what is actually being narrowed,
otherwise we obviously won't be able to produce the right narrow. But it also means
that the critical tree-level operations `And` and `Or` need to be "global" in
the sense that they should apply only to a top-level name, *not* to a specific
attribute chain.

## This diff

From the considerations above, I have concluded that to implement attribute narrowing
within the current infrastructure, I need the full `NarrowOp` tree that represents
a complex narrow op to be distinct from a "simple" `NarrowOp` that represents a specific
narrow, because the former is at the level of an entire `TypeInfo` whereas the latter
is a `Type` level operation that applies to a specific spot within the `TypeInfo`, and
the inner operations also need to eventually have "tags" in the data structure that
allow us to figure out which attribute access chain is relevant.

This diff is the approach I came up with: split `NarrowOp` into `AtomicNarrowOp` representing
a specific narrow that applies to a `Type` vs the `NarrowOp` type that represent a
tree over boolean ops and applies to an entire `TypeInfo`. Fully supporting attribute narrows
will require that the `NarrowOp::Atomic` case, which as of this diff just wraps an `AtomicNarrowOp`,
also include some sort of optional attribute chain.

Reviewed By: rchen152

Differential Revision: D72541244

fbshipit-source-id: 62ca8cd83e7bf6212f2d043066a691b552bb560f