Introduce a `NarrowingSubject` enum
Summary:
In order to thread attribute narrowing information through bindings, we need to
be able to collect and work with attribute chain narrows and not just top-level
narrows, which means the existing `expr_to_names` function needs to be extended
to be able to work with attribute chains as the "subject" of a narrow.

This diff introduces an enum with no new functionality, I'll work on adding attribute
accesses next.

Unlike in an earlier version of this work, the new enum is private because we
actually need to collect all of the subjects with the same base name into a single
NarrowOps, which means that the enums will always be decomposed in `NarrowOp::from_expr`.

Reviewed By: ndmitchell

Differential Revision: D72564335

fbshipit-source-id: a52ff5fb17926062ca19a6d06977bb0153a0bc99