Move config overriding into main.rs
Summary:
Over this and the next diff, I'm going to be moving the functionality for loading/overriding a config into `main.rs` (specifically into the config finder functionality). I'm doing this because:
* we can cache already loaded/overridden configs, and return those instead of re-parsing/overriding every time we load
* we can avoid re-emitting an error message every time we deal with it (which in project checking mode, will happen on every file that will be type checked).

It also feels right to me that we'd do the overriding in the `config_finder` function, rather than as a separate step once we've loaded a config everywhere we do that (check.rs and lsp.rs). In the future, once this function gets moved so that we can share it with lsp.rs, having this just take a function that overrides the config with values from args will make things easier.

Reviewed By: grievejia

Differential Revision: D72436431

fbshipit-source-id: 655e8b2d2a174efa687e856a378742b5680d52c7