Fix attribute lookup bug caused by attempt to mutate type params on missing default
Summary: When we found a type parameter with a missing default, we tried to mutate it to add a default of Any. This caused the QuantifiedInfo stored in a class's TParams to no longer match the (unmutated) QuantifiedInfo in class fields that refer to the type parameter. Luckily, we don't actually need to mutate anything because we always fall back to Any in the absence of a default anyway. So I just deleted the code that did the mutating.

Reviewed By: ndmitchell

Differential Revision: D72617536

fbshipit-source-id: 8c1d5c29518ad287ece772d24d6cfd2aee5b51f1