Document that issubclass on tuple does not work
Summary:
It is not convenient to try to make this work with attribute narrowing, but it
turns out the logic is already broken.

Better to just remove incorrect support and make someone fix-forward than to
try to maintain logic that does not work properly.

Note that `issubclass` is much less common than `isinstance`, so this is
much lower-priority anyway.

Reviewed By: ndmitchell, rchen152

Differential Revision: D72579809

fbshipit-source-id: 7191d26af1cf2b45d2dec235fd86a9658aa08cb1