Remove buggy `issubclass(_, <tuple>)` support
Summary:
As documented in our new test, it doesn't work properly in the first place.

Better to not even attempt to support it, which leads to predictable behavior
and an error message that indicates the real problem.

This helps to unblock attribute narrowing because the buggy implementation was
also relying on constructing trees of atomic narrow ops from a single
atomic narrow op, which won't play well with attribute narrowing (which requires
us to treat atomic ops as `Type`-level but trees as `TypeInfo`-level)

Reviewed By: ndmitchell, yangdanny97

Differential Revision: D72579811

fbshipit-source-id: e7e4c0281bf3b88ad721f7720b56f288a9e644ab