Add an attribute tag to `NarrowOp::Atomic`
Summary:
The purpose of splitting `NarrowOp` and `AtomicNarrowOp` was to
enable us to tag the atomic operations with an optional attribue chain
to which each operation applies.

This diff adds the plumbing to do that; an attribute is an optional
`Vec1` of names defining the access chain.

At this stage it is still always None because all atomic ops are
currently applied to top-level names; in the solver-side logic we
ignore any narrows that apply to non-None attributes for now since we
are not yet ready to perform these narrows, but we are ready to start
creating them (and we want the intermediate diffs to be release-safe).

Reviewed By: ndmitchell

Differential Revision: D72579806

fbshipit-source-id: 2c04e33e4a4d35192a4cdce7b7f65bd6927a295d